backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
openai==1.57.2
google-generativeai==0.8.3
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
